add G23 keyboard
