--- keyboards/shameless_winkeyless_prime/thirty_one_key/info.json.numbers
@@ -1,38 +1,0 @@
-{
-    "version": 1,
-    "notes": "",
-    "documentation":"\"This file is a QMK Configurator export. You can import this at <https://config.qmk.fm>. It can also be used directly with QMK's source code.\n\nTo setup your QMK environment check out the tutorial: <https://docs.qmk.fm/#/newbs>\n\nYou can convert this file to a keymap.c using this command: `qmk json2c {keymap}`\n\nYou can compile this keymap using this command: `qmk compile {keymap}`\"\n",
-    "keyboard": "bonksalot/nine_key",
-    "keymap": "nine",
-    "keyboard_name": "nine_key",
-    "url": "https://www.hissing-skuz.de",
-    "maintainer": "filmon",
-    "width": 3,
-    "height": 3,
-    "layout": "LAYOUT",
-    "layers": [
-    [
-      "KC_7",
-      "KC_8",
-      "KC_9",
-      "KC_4",
-      "LT(1,KC_5)",
-      "KC_6",
-      "KC_1",
-      "KC_2",
-      "KC_3"
-    ],
-    [
-      "RGB_RMOD",
-      "RGB_VAI",
-      "RGB_MOD",
-      "RGB_HUI",
-      "KC_TRNS",
-      "RGB_SAI",
-      "RGB_HUD",
-      "RGB_VAD",
-      "RGB_SAD"
-    ]
-  ],
-  "author": ""
-}
