--- keyboards/shameless_winkeyless_prime/thirty_one_key/keymaps/nine/keymap.c
@@ -1,13 +1,0 @@
-#include QMK_KEYBOARD_H
-
-/* THIS FILE WAS GENERATED!
- *
- * This file was generated by qmk json2c. You may or may not want to
- * edit it directly.
- */
-
-const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
-	[0] = LAYOUT(KC_ESC, LT(1, RGB_TOG), KC_BSPC, KC_WH_L, KC_UP, KC_WH_R, KC_LEFT, KC_DOWN, KC_RIGHT),
-	[1] = LAYOUT(RGB_RMOD, KC_NO, RGB_MOD, RGB_HUI, RGB_VAI, RGB_SAI, RGB_HUD, RGB_VAD, RGB_SAD)
-};
-
