--- keyboards/shameless_winkeyless_prime/thirty_one_key/keymaps/via/config.h
@@ -1,1 +1,0 @@
-#define DYNAMIC_KEYMAP_LAYER_COUNT 3
