--- keyboards/shameless_winkeyless_prime/thirty_one_key/keymaps/via/rules.mk
@@ -1,2 +1,0 @@
-VIA_ENABLE = yes
-LTO_ENABLE = yes
