--- keyboards/shameless_winkeyless_prime/thirty_one_key/readme.md
@@ -1,23 +1,0 @@
-# B.mini
-
-A 75% keyboard with RGB underglow. 
-
-QMK Firmware is not the official firmware for the B.mini. 
-
-* Keyboard Maintainer: QMK Community
-* Hardware Supported: B.mini PCB
-* Hardware Availability: <http://winkeyless.kr/product/b-mini-x2-pcb/>
-
-Make example for this keyboard (after setting up your build environment):
-
-    make winkeyless/bmini:default
-
-Flashing example for this keyboard ([after setting up the bootloadHID flashing environment](https://docs.qmk.fm/#/flashing_bootloadhid))
-
-    make winkeyless/bmini:default:flash
-
-This firmware can be flashed onto the KBDFans KBD75 Round 1 and Round 2 Bootmapper Client boards sold during the 2017 and early 2018 timeframe. 
-
-**Reset Key**: Hold down the bottom left most key located at `K00`, commonly programmed as Left Control while plugging in the keyboard.
-
-See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
