+++ keyboards/shameless_winkeyless_prime/twenty_four_key/keymaps/default/info.json
@@ -1,0 +1,116 @@
+{
+    "version": 1,
+    "notes": "This keymap is for a 9 key number pad",
+    "documentation":"\"This file is a QMK Configurator export. You can import this at <https://config.qmk.fm>. It can also be used directly with QMK's source code.\n\nTo setup your QMK environment check out the tutorial: <https://docs.qmk.fm/#/newbs>\n\nYou can convert this file to a keymap.c using this command: `qmk json2c {keymap}`\n\nYou can compile this keymap using this command: `qmk compile {keymap}`\"\n",
+    "keyboard": "shameless_winkeyless_prime/nine_key",
+    "keymap": "default",
+    "keyboard_name": "nine_key",
+    "url": "https://www.hissing-skuz.de",
+    "maintainer": "filmon",
+    "width": 8,
+    "height": 6,
+    "layout": "LAYOUT",
+    "layers": [
+    [
+      "KC_NO",
+      "BL_STEP",
+      "KC_1",
+      "KC_2",
+      "KC_3",
+      "KC_NO",
+      "KC_MS_LEFT",
+      "KC_MS_RIGHT",
+      "KC_ESCAPE",
+      "KC(L'ä')",
+      "KC_6",
+      "KC_7",
+      "KC_A",
+      "KC_BSPACE",
+      "KC_MS_UP",
+      "KC_MS_DOWN",
+      "KC_Q",
+      "KC_W",
+      "KC_E",
+      "KC_H",
+      "KC_I",
+      "KC_J",
+      "KC_MS_WH_DOWN",
+      "KC_MS_WH_UP",
+      "KC_A",
+      "KC_S",
+      "KC_D",
+      "KC_P",
+      "KC_UP",
+      "KC_ENTER",
+      "KC_MS_BTN1",
+      "KC_MS_BTN2",
+      "KC_LCTRL",
+      "KC_W",
+      "KC_X",
+      "KC_LEFT",
+      "KC_DOWN",
+      "KC_RIGHT",
+      "KC_MS_BTN3",
+      "KC_NO",
+      "KC_NO",
+      "KC_NO",
+      "KC_NO",
+      "KC_NO",
+      "KC_NO",
+      "KC_NO",
+      "KC_NO",
+      "KC_NO"
+    ],
+    [
+      "RGB_RMOD",
+      "RGB_MOD",
+      "RGB_HUI",
+      "RGB_VAI",
+      "KC_TRNS",
+      "RGB_SAI",
+      "RGB_HUD",
+      "RGB_VAD",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "RGB_SAD"
+    ]
+  ],
+  "author": ""
+}
