+++ keyboards/shameless_winkeyless_prime/twenty_four_key/keymaps/default/keymap.c
@@ -1,0 +1,13 @@
+#include QMK_KEYBOARD_H
+
+/* THIS FILE WAS GENERATED!
+ *
+ * This file was generated by qmk json2c. You may or may not want to
+ * edit it directly.
+ */
+
+const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
+	[0] = LAYOUT(KC_NO, BL_STEP, KC_1, KC_2, KC_3, KC_NO, KC_MS_LEFT, KC_MS_RIGHT, KC_ESCAPE, KC(L'ä'), KC_6, KC_7, KC_A, KC_BSPACE, KC_MS_UP, KC_MS_DOWN, KC_Q, KC_W, KC_E, KC_H, KC_I, KC_J, KC_MS_WH_DOWN, KC_MS_WH_UP, KC_A, KC_S, KC_D, KC_P, KC_UP, KC_ENTER, KC_MS_BTN1, KC_MS_BTN2, KC_LCTRL, KC_W, KC_X, KC_LEFT, KC_DOWN, KC_RIGHT, KC_MS_BTN3, KC_NO, KC_NO, KC_NO, KC_NO, KC_NO, KC_NO, KC_NO, KC_NO, KC_NO),
+	[1] = LAYOUT(RGB_RMOD, RGB_MOD, RGB_HUI, RGB_VAI, KC_TRNS, RGB_SAI, RGB_HUD, RGB_VAD, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS, RGB_SAD)
+};
+
