+++ keyboards/shameless_winkeyless_prime/twenty_four_key/rules.mk
@@ -13,10 +13,11 @@ BOOTLOADER = bootloadHID
 
 # build options
 BOOTMAGIC_ENABLE = full
-MOUSEKEY_ENABLE = no
+MOUSEKEY_ENABLE = yes
 EXTRAKEY_ENABLE = yes
 CONSOLE_ENABLE = no
 COMMAND_ENABLE = yes
 BACKLIGHT_ENABLE = yes
-RGBLIGHT_ENABLE = yes
+RGBLIGHT_ENABLE = no
 WS2812_DRIVER = i2c
+UNICODE_ENABLE = yes
